Few corrections

+++ src/Particles/nvidiaParticles.h
@@ -33,7 +33,7 @@ public:
     std::vector< thrust::device_vector<double> > nvidiaMomentum;
     thrust::device_vector<double> nvidiaWeight;
     thrust::device_vector<short>  nvidiaCharge;
-    thrust::device_vector<short>  nvidiaChi;
+    thrust::device_vector<double>  nvidiaChi;
 
     //! cell_keys of the particle
     thrust::device_vector<int> nvidia_cell_keys;
@@ -50,7 +50,7 @@ public:
     short * getPtrCharge() override {
         return thrust::raw_pointer_cast( nvidiaCharge.data() );
     };
-    short * getPtrChi() override {
+    double * getPtrChi() override {
         return thrust::raw_pointer_cast( nvidiaChi.data() );
     };
     int * getPtrCellKeys() override {
